post_url converted to post_text
The challenge does not need a URL but instead wants a body for a ton of text.  Ran into an error with string and found out it is limited to 255 characters and you have to change it to TEXT to allow more.  I then updated comments_text to TEXT as well to allow for more room.  The models should meet the requirements now with the routes needing to be updated as more things get bolted on.