cleaned up a little hmtl and more #if
started to try and make the html match the requirements by dropping the comments counter and making the title of the post link to the single post page.  Also added a few #if's to hide the js in single post so if you are not logged it it will not display the box or expect the boxes js to be there.